feat(data): Build foundational data layer for new features

+++ app/src/main/java/com/example/myapplication/data/BehaviorEventDao.kt
@@ -17,5 +17,5 @@ interface BehaviorEventDao {
     fun getAllBehaviorEvents(): LiveData<List<BehaviorEvent>>
 
     @Query("SELECT * FROM behavior_events WHERE studentId = :studentId ORDER BY timestamp DESC LIMIT :limit")
-    fun getRecentBehaviorEventsForStudent(studentId: Long, limit: Int): LiveData<List<BehaviorEvent>>
+    fun getRecentBehaviorEventsForStudent(studentId: Int, limit: Int): LiveData<List<BehaviorEvent>>
 }
